pyproject.toml
[project]
name = "agent-labs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.7"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.112",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
]
